Functioning AI!!!!
Holy shit what an update!
biggest news is that i got the AI to learn and update the database reliably!
The bad news is that there are still two known bugs:
- if the AI wins it cannot update the final move (because it is an "O" and not a Zero or Int). This can be fixed by checking to see if the position .Equals("X") || ("O"); else learn!

-The Win() validation is broken. i need to fix this by just getting rid of the variable ninetyDegrees (which is the board flipped ninety degrees) and just using the string array gameBoard.

those are the two bugs, heres the update!!!

GameBoardsController
updated the final piece of code so that i dont get the multi ID tracking error, i do know why i was getting it so the teacher might be interested in knowing. code for that is still below
-Got rid of UpdateChoices, might put it back in with the new code and see if it runs, it is cleaner.
Game(){
- minor changes to the win() method, but it will need to be revamped as stated earlier
- UpdateAI will be deleted because the context cannot work outside of its scope and therefore is left in the Controller.
}
GameBoard(){
- i fixed the Equals method, not the most efficient thing in the world but i'll be damned if it doesn't work.
}