bug fixes and other small tweaks
Before i left this on the master i made some small noticable changes:
Game(){
added a cats game feature which accuratly tells if the game is over or not win or not
}
GameBoard(){
Still trying to make the AI learn but its being a pain in my rear end, might have to make a new controller for that.
}

I also just made it able to run on the master after the merge.